+++ packages/connectors/gcp/jest.config.js
@@ -1,0 +1,5 @@
+const commonConfig = require('@elbwalker/jest/node.config');
+
+const config = {};
+
+module.exports = { ...commonConfig, ...config };