+++ packages/connectors/gcp/src/types/index.ts
@@ -1,0 +1,6 @@
+import type { Request } from '@elbwalker/types';
+
+export interface HttpFunction {
+  anonymizeIp?: boolean;
+  mapping?: Record<string, keyof Request.Context | false>;
+}