+++ packages/clients/walkerjs/CHANGELOG.md
@@ -1,12 +1,20 @@
 # @elbwalker/walker.js
 
+## 3.1.1
+
+### Patch Changes
+
+- Updated dependencies [5c33b75]
+- Updated dependencies [d3735ad]
+  - @elbwalker/utils@3.5.1
+
 ## 3.1.0
 
 ### Minor Changes
 
 - 5a2c0b0: Prefer an events consent state
   [#431](https://github.com/elbwalker/walkerOS/issues/431)
-- 934c5a7: Data Policy [#447]https://(github.com/elbwalker/walkerOS/issues/447)
+- 934c5a7: Data Policy [#447](https://github.com/elbwalker/walkerOS/issues/447)
 
 ### Patch Changes
 
@@ -26,7 +34,7 @@
 ### Patch Changes
 
 - cc2c041: elb with custom elbLayer support
-  [#419[(https://github.com/elbwalker/walkerOS/issues/419)
+  [#419](https://github.com/elbwalker/walkerOS/issues/419)
 - Updated dependencies [cc2c041]
   - @elbwalker/utils@3.4.1