+++ packages/destinations/web/piwikpro/CHANGELOG.md
@@ -1,10 +1,19 @@
 # @elbwalker/destination-web-piwikpro
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [5c33b75]
+- Updated dependencies [d3735ad]
+  - @elbwalker/utils@3.5.1
+  - @elbwalker/walker.js@3.1.1
+
 ## 1.3.0
 
 ### Minor Changes
 
-- 934c5a7: Data Policy [#447]https://(github.com/elbwalker/walkerOS/issues/447)
+- 934c5a7: Data Policy [#447](https://github.com/elbwalker/walkerOS/issues/447)
 
 ### Patch Changes